requirements.txt
opencv-python<=4.10.0.84
shapely>=2.0.0
tqdm>=4.48.2
pillow>=8.2.0
pybboxes==0.1.6
pyyaml
fire
terminaltables
requests
click
opencv-python<=4.10.0.84
globox
torch
torchvision
ultralytics
